comfyapp-dev/phase1-mvp/backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pillow>=10.0.0
diffusers>=0.24.0
transformers>=4.35.0
torch>=2.1.0
torchvision>=0.16.0
accelerate>=0.24.0
numpy>=1.24.0
requests>=2.31.0